requirements.txt
torch
torchvision
numpy
matplotlib
pillow
scikit-learn